backend/requirements.txt
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.4
django-cors-headers==3.1.0
django-cors-headers==3.0.1
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
oauthlib==3.1.0
psycopg2==2.8.3
PyJWT==1.7.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
pillow==8.0.1
django-resized==0.3.11
django-ses==1.0.3
reportlab==3.5.66
python-jose
python3-dev
cryptography
changelogs